backend/services/embedding/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers>=3.0,<4.0
numpy>=1.24,<2
pydantic
python-dotenv
